Renamed all to ViewRegister
