+++ TinYard/Extensions/ViewController/API/Interfaces/IViewRegister.cs
@@ -1,6 +1,6 @@
 ﻿namespace TinYard.Extensions.ViewController.API.Interfaces
 {
-    public interface IViewController
+    public interface IViewRegister
     {
 
     }
